modified:   app.py 	modified:   cache.py 	modified:   document_qa.py 	modified:   model_state.py 	modified:   multimodal.py

+++ app.py
@@ -13,7 +13,6 @@ import global_var
 from cache import (
     init_cache,
     load_cache,
-    create_new_cache,
     Cache,
     CacheConst,
     change_running_cache,
@@ -31,7 +30,14 @@ from document_qa import document_qa_fn
 import numpy as np
 from PIL import Image
 from multimodal import multimodal_chat
-from model_state import LLMState, LLMStateConst, MLLMState, MLLMStateConst
+from model_state import (
+    LLMState,
+    LLMStateConst,
+    MLLMState,
+    MLLMStateConst,
+    EMBStateConst,
+    EMBState,
+)
 from tools import (
     RetrievalState,
     ToolsState,
@@ -52,6 +58,9 @@ def _init_state_vars():
     global_var.set_global_value("websearch_state", WebSearchState())
     global_var.set_global_value("tools_state", ToolsState())
     global_var.set_global_value("chat_history", [])
+    global_var.set_global_value("pdf_llm_state", LLMState())
+    global_var.set_global_value("emb_state", EMBState())
+
 
 def wav2txt(path: str, lang: str = "chinese"):
     model = wenet.load_model(lang)
@@ -116,13 +125,17 @@ def chat_with_document(filepath: str, question: str, chat_history: list):
         yield chat_history, None
     yield chat_history, gr.Textbox(interactive=True)
 
+
 def on_select(evt: gr.SelectData):
-    cache:Cache = global_var.get_global_value('cache')
+    cache: Cache = global_var.get_global_value("cache")
     gr.Info(f"已在PDF阅读器中打开“{evt.value[0]}”，切换到“PDF文档问答”栏以查看详情")
-    return str(cache.cached_files_dir/evt.value[0])
+    return str(cache.cached_files_dir / evt.value[0])
+
 
 def delete_all_files():
     cache = load_cache()
+    if cache is None:
+        raise gr.Error("请先建立知识库")
     cache.clear_all()
     gr.Info(f"所有文章都已被删除")
     return (
@@ -137,9 +150,11 @@ def delete_all_files():
 
 def delete_files(filenames: list[str]):
     cache = load_cache()
+    if cache is None:
+        raise gr.Error("请先建立知识库")
     if isinstance(filenames, list) == False:
         gr.Info("请选择要删除的文章")
-        filenames=[]
+        filenames = []
     for filename in filenames:
         cache.delete_file(filename)
         gr.Info(f"文件 ({filename})已经被删除")
@@ -152,7 +167,9 @@ def delete_files(filenames: list[str]):
 def upload(files: list[str]):
     cache_config = cast(dict[str, str], global_var.get_global_value("cache_config"))
     cache = load_cache()
-    logger.info(files)
+    if cache is None:
+        raise gr.Error("请先建立知识库！")
+    logger.info(f"upload files:{files}")
     for filepath in files:
         cache.cache_file(filepath, save_local=True)
         gr.Info(f"文件 {Path(filepath).name} 上传成功!")
@@ -162,32 +179,45 @@ def upload(files: list[str]):
     ), [[i] for i in cache.all_files]
 
 
-def create_cache(namespace: str, emb_model_name: str):
+def create_cache(namespace: str):
     valid = True
+    emb_state: EMBState = global_var.get_global_value("emb_state")
     if isinstance(namespace, str) == False or namespace == "":
         gr.Info("请输入知识库的名字")
         valid = False
-    if isinstance(emb_model_name, str) == False or emb_model_name == "":
+    if isinstance(emb_state.model, str) == False or emb_state.model == "":
         gr.Info("请选择Embbeding模型")
         valid = False
     if valid == False:
-        return gr.update(), gr.update(), gr.update()
+        return gr.update(), gr.update(), gr.update(), gr.update(), gr.update()
+    if isinstance(emb_state.api_key,str)==False or emb_state.api_key=='':
+        emb_state.api_key = None
+    if isinstance(emb_state.base_url,str)==False or emb_state.base_url=='':
+        emb_state.base_url = None
     cache_lst: list[str] = global_var.get_global_value("cache_lst")
     if namespace in cache_lst:
         gr.Warning(f"知识库“{namespace}”已经存在，故无法创建")
         cache = load_cache(namespace)
     else:
-        cache = Cache(namespace=namespace, emb_model_name=emb_model_name)
+        print(emb_state.model_dump())
+        cache = Cache(
+            namespace=namespace,
+            emb_model_name=emb_state.model,
+            emb_api_key=emb_state.api_key,
+            emb_base_url=emb_state.base_url,
+        )
         gr.Info(f"新的知识库“{namespace}”创建成功")
     return (
         gr.update(value=None),
         gr.update(value=None),
         gr.update(value=None, choices=cache_lst),
+        gr.update(value=None),
+        gr.update(value=None),
     )
 
 
 def change_cache(namespace: str):
-    if isinstance(namespace, str)==False:
+    if isinstance(namespace, str) == False:
         raise gr.Error("请选择知识库！")
     cache = change_running_cache(namespace)
     gr.Info("成功更改本地知识库设置")
@@ -201,11 +231,13 @@ def change_cache(namespace: str):
 
 def delete_cache(namespace: str):
     cache: Cache = global_var.get_global_value("cache")
+    if cache is None:
+        raise gr.Error("请先建立知识库！")
     if cache.namespace == namespace:
         gr.Info("无法删除正在使用的本地知识库")
         return gr.update()
-    elif isinstance(namespace,str):
-        cache = load_cache(namespace=namespace)
+    elif isinstance(namespace, str):
+        cache = cast(Cache, load_cache(namespace=namespace))
         cache.delete_cache()
         cache_lst: list[str] = global_var.get_global_value("cache_lst")
         gr.Info(f"成功删除本地知识库{namespace}")
@@ -242,7 +274,6 @@ def get_timestamp():
     return _timestamp
 
 
-
 def create_ui():
     def _state_change():
         global _state
@@ -260,17 +291,28 @@ def create_ui():
             if _state["name"] == "warning":
                 gr.Warning(cast(str, _state["message"]))
         return {modal: Modal(visible=False), modalMsg: ""}
-    
+
     def load_demo():
-        cache: Cache = global_var.get_global_value("cache")
-        cache_lst:list[str] = global_var.get_global_value('cache_lst')
-        # chat_history: list = gl
+        cache: Cache|None = global_var.get_global_value("cache")
+        cache_lst: list[str] = global_var.get_global_value("cache_lst")
+        mllm_state: MLLMState = global_var.get_global_value("mllm_state")
+        llm_state: LLMState = global_var.get_global_value("llm_state")
+        pdf_llm_state: LLMState = global_var.get_global_value("pdf_llm_state")
         return {
-            dstCachedPapers:[[i] for i in cache.all_files],
-            cacheFilesDdl: gr.update(value=None, choices=cache.all_files),
-            cacheNameDdl: gr.update(value=cache.namespace, choices=cache_lst),
-            currEmbTxt: cache.emb_model_name,
-            currNamespaceTxt: cache.namespace
+            dstCachedPapers: [[i] for i in cache.all_files] if cache else [],
+            cacheFilesDdl: gr.update(value=None, choices=cache.all_files) if cache else [],
+            cacheNameDdl: gr.update(value=cache.namespace, choices=cache_lst) if cache else None,
+            currEmbTxt: cache.emb_model_name if cache else None,
+            currNamespaceTxt: cache.namespace if cache else None,
+            mllmDdl: mllm_state.model,
+            mllmApikeyDdl: mllm_state.api_key,
+            mllmBaseurlTxt: mllm_state.base_url,
+            llmDdl: llm_state.model,
+            llmApikeyDdl: llm_state.api_key,
+            llmBaseurlTxt: llm_state.base_url,
+            pdfLlmDdl: pdf_llm_state.model,
+            pdfLlmApikeyDdl: pdf_llm_state.api_key,
+            pdfLlmBaseurlTxt: pdf_llm_state.base_url,
         }
 
     with gr.Blocks(title="SciAgent", theme="soft") as demo:
@@ -322,7 +364,7 @@ def create_ui():
                         llmDdl = gr.Dropdown(
                             choices=SUPPORT_LLMS,  # type: ignore
                             value=llm_state.model,
-                            label="大语言模型ID",
+                            label="模型ID",
                         )
                         llmApikeyDdl = gr.Textbox(
                             label="模型api-key",
@@ -425,32 +467,28 @@ def create_ui():
                             ],
                         )
 
-        with gr.Tab(label="本地知识库") as localKb:
-            cache = load_cache()
-            cache_config = cast(
-                dict[str, Any], global_var.get_global_value("cache_config")
-            )
-            cache_lst = cast(list[str], global_var.get_global_value("cache_lst"))
+        with gr.Tab(label="本地知识库"):
             with gr.Row():
                 with gr.Column(scale=3):
                     dstCachedPapers = gr.Dataset(
                         components=[gr.Textbox(visible=False)],
                         label=f"已存储的文章",
-                        samples=[[i] for i in cache.all_files],
+                        # samples=[[i] for i in cache.all_files],
                     )
                 with gr.Column(scale=1):
                     currNamespaceTxt = gr.Textbox(
-                        label="当前知识库", value=cache.namespace, interactive=False
+                        label="当前知识库", 
+                        # value=cache.namespace, 
+                        interactive=False
                     )
                     currEmbTxt = gr.Textbox(
                         label="当前知识库使用的Embedding模型",
-                        value=cache.emb_model_name,
+                        # value=cache.emb_model_name,
                         interactive=False,
                     )
                     with gr.Accordion(label="文件管理"):
                         cacheFilesDdl = gr.Dropdown(
                             label="选择要从本地知识库中删除的文章",
-                            choices=cache.all_files,  # type:ignore
                             multiselect=True,
                         )
                         uploadFileBot = gr.File(
@@ -476,30 +514,51 @@ def create_ui():
                     with gr.Accordion(label="新建知识库", open=False):
                         newNamespaceTxt = gr.Textbox(
                             label="知识库名称",
-                            # info="名称中不能包含下划线"
+                            info="名称应为3-63个字母或者数字组成的字符串"
                         )
                         newEmbDdl = gr.Dropdown(
-                            choices=cast(
-                                list[str | int | float | tuple[str, str | int | float]]
-                                | None,
-                                SUPPORT_EMBS,
-                            ),
+                            choices=EMBStateConst.EMB_CHOICES,  # type:ignore
                             multiselect=False,
                             label="Embedding模型",
                         )
+                        newEmbApiKeyDdl = gr.Textbox(
+                            label="模型api-key",
+                            value=EMBStateConst.DEFAULT_API_KEY,
+                            type="password",
+                        )
+                        newEmbBaseurlTxt = gr.Textbox(
+                            label="模型baseurl",
+                            value=EMBStateConst.DEFAULT_BASE_URL,
+                            info="如使用Openai模型此栏请留空",
+                        )
                         newCacheButton = gr.Button(value="新建本地知识库")
+                        gr.on(
+                            [
+                                newEmbDdl.change,
+                                newEmbApiKeyDdl.change,
+                                newEmbBaseurlTxt.change,
+                            ],
+                            lambda model, api_key, base_url: state.change_state(
+                                "emb_state",
+                                model=model,
+                                api_key=api_key,
+                                base_url=base_url,
+                            ),
+                            inputs=[newEmbDdl, newEmbApiKeyDdl, newEmbBaseurlTxt],
+                        )
+
                     with gr.Accordion(label="更改知识库", open=False):
                         cacheNameDdl = gr.Dropdown(
-                            value=cache.namespace,
-                            choices=cache_lst,  # type:ignore
+                            # value=cache.namespace,
+                            # choices=cache_lst,  # type:ignore
                             label="选择知识库",
                         )
                         changeCacheBtn = gr.Button(value="更改为选中的本地知识库")
                         deleteCacheBtn = gr.Button(value="删除选中的本地知识库")
                     newCacheButton.click(
                         create_cache,
-                        inputs=[newNamespaceTxt, newEmbDdl],
-                        outputs=[newNamespaceTxt, newEmbDdl, cacheNameDdl],
+                        inputs=[newNamespaceTxt],
+                        outputs=[newNamespaceTxt, newEmbDdl, cacheNameDdl, newEmbApiKeyDdl, newEmbBaseurlTxt],
                     )
                     changeCacheBtn.click(
                         change_cache,
@@ -515,13 +574,46 @@ def create_ui():
                         delete_cache, inputs=[cacheNameDdl], outputs=[cacheNameDdl]
                     )
 
-
         # with gr.Tab(label="工作台"):
         #     pass
         with gr.Tab(label="PDF文档问答"):
             with gr.Row():
                 with gr.Column():
-                    pdfBox = PDF(label="PDF文档", height=700)
+                    with gr.Accordion("模型设置"):
+                        pdfLlmDdl = gr.Dropdown(
+                            choices=LLMStateConst.LLM_CHOICES,  # type:ignore
+                            value=LLMStateConst.DEFAULT_LLM,
+                            label="模型ID",
+                        )
+                        pdfLlmApikeyDdl = gr.Textbox(
+                            label="模型api-key",
+                            type="password",
+                            value=LLMStateConst.DEFAULT_API_KEY,
+                        )
+                        pdfLlmBaseurlTxt = gr.Textbox(
+                            label="模型baseurl",
+                            info="如使用Openai模型此栏请留空",
+                            value=LLMStateConst.DEFAULT_BASE_URL,
+                        )
+                    pdfBox = PDF(
+                        label="PDF文档",
+                        height=700,
+                        info="模型上下文长度有限，请不要传入太长的文章。",
+                    )
+                    gr.on(
+                        [
+                            pdfLlmDdl.change,
+                            pdfLlmApikeyDdl.change,
+                            pdfLlmBaseurlTxt.change,
+                        ],
+                        lambda model, api_key, base_url: state.change_state(
+                            "pdf_llm_state",
+                            model=model,
+                            api_key=api_key,
+                            base_url=base_url,
+                        ),
+                        inputs=[pdfLlmDdl, pdfLlmApikeyDdl, pdfLlmBaseurlTxt],
+                    )
                 with gr.Column():
                     docChatbot = gr.Chatbot(label="问答记录", height=800)
                     docTxtbot = gr.Textbox(
@@ -581,6 +673,7 @@ def create_ui():
                                 mllmApikeyDdl.change,
                                 mllmBaseurlTxt.change,
                                 mllmMaxTokens.change,
+                                demo.load,
                             ],
                             lambda model, api_key, base_url, max_tokens: state.change_state(
                                 "mllm_state",
@@ -644,13 +737,24 @@ def create_ui():
         )
         demo.load(
             fn=load_demo,
-            outputs=[dstCachedPapers, cacheFilesDdl, cacheNameDdl, currEmbTxt, currNamespaceTxt],
+            outputs=[
+                llmDdl,
+                llmApikeyDdl,
+                llmBaseurlTxt,
+                dstCachedPapers,
+                cacheFilesDdl,
+                cacheNameDdl,
+                currEmbTxt,
+                currNamespaceTxt,
+                pdfLlmDdl,
+                pdfLlmApikeyDdl,
+                pdfLlmBaseurlTxt,
+                mllmDdl,
+                mllmApikeyDdl,
+                mllmBaseurlTxt
+            ],
         )
-        dstCachedPapers.select(
-            on_select,
-            outputs=pdfBox,
-            scroll_to_output=True
-        )
+        dstCachedPapers.select(on_select, outputs=pdfBox, scroll_to_output=True)
     return demo
